refactor dynamic state description provide to commons
